moves isolated AGSPopupAttachmentType+Icon to caller file
