+++ data-collection/data-collection/View Controllers/Rich Popup View Controllers/Rich Popup View Controller/RichPopupAttachmentsViewController/RichPopupAttachmentsViewController+UITableViewDataSource.swift
@@ -95,3 +95,23 @@ extension RichPopupAttachmentsViewController /* UITableViewDataSource */ {
         }
     }
 }
+
+// MARK: - Attachment Type
+
+extension AGSPopupAttachmentType {
+    
+    var icon: UIImage {
+        switch self {
+        case .image:
+            return #imageLiteral(resourceName: "AttachmentImage")
+        case .video:
+            return #imageLiteral(resourceName: "AttachmentVideo")
+        case .document:
+            return #imageLiteral(resourceName: "AttachmentDocument")
+        case .other:
+            return #imageLiteral(resourceName: "AttachmentOther")
+        @unknown default:
+            fatalError("Unsupported case \(self).")
+        }
+    }
+}
